backend_rag/requirements.txt
fastapi[standard]
uvicorn
chromadb
sentence-transformers
pymupdf
python-docx
rank_bm25
jieba
pydantic
